backend/src/scripts/python/requirements.txt
scikit-learn>=1.0.2
geopandas>=0.10.2
spacy>=3.4.0
pandas>=1.5.0
numpy>=1.20.0
shapely>=1.8.0
rtree>=1.0.0
matplotlib>=3.5.0
tensorflow>=2.9.0
psycopg2-binary>=2.9.3 